Convert CheckoutPending to TS
